ignore String() method to fix coverage
