+++ codecov.yml
@@ -1,0 +1,3 @@
+ignore:
+  - "*_string.go"
+  - "*/*_string.go"
