+++ contrib/img_hash.go
@@ -72,16 +72,6 @@ const (
 	BlockMeanHashModeDefault = BlockMeanHashMode0
 )
 
-func (c BlockMeanHashMode) String() string {
-	switch c {
-	case BlockMeanHashMode0:
-		return "block-mean-hash-mode0"
-	case BlockMeanHashMode1:
-		return "block-mean-hash-mode1"
-	}
-	return ""
-}
-
 // Compute computes hash of the input image using BlockMeanHash.
 //
 // For further information, see:
