+++ core.go
@@ -68,32 +68,6 @@ const (
 	MatTypeCV8UC4 = MatTypeCV8U + MatChannels4
 )
 
-func (c MatType) String() string {
-	switch c {
-	case MatTypeCV8U:
-		return "CV8U"
-	case MatTypeCV16U:
-		return "CV16U"
-	case MatTypeCV16S:
-		return "CV16S"
-	case MatTypeCV16SC2:
-		return "CV16SC2"
-	case MatTypeCV32S:
-		return "CV32S"
-	case MatTypeCV32F:
-		return "CV32F"
-	case MatTypeCV64F:
-		return "CV64F"
-	case MatTypeCV8UC2:
-		return "CV8UC2"
-	case MatTypeCV8UC3:
-		return "CV8UC3"
-	case MatTypeCV8UC4:
-		return "CV8UC4"
-	}
-	return ""
-}
-
 // CompareType is used for Compare operations to indicate which kind of
 // comparison to use.
 type CompareType int
@@ -117,24 +91,6 @@ const (
 	// CompareNE src1 is unequal to src2.
 	CompareNE = 5
 )
-
-func (c CompareType) String() string {
-	switch c {
-	case CompareEQ:
-		return "eq"
-	case CompareGT:
-		return "gt"
-	case CompareGE:
-		return "ge"
-	case CompareLT:
-		return "lt"
-	case CompareLE:
-		return "le"
-	case CompareNE:
-		return "ne"
-	}
-	return ""
-}
 
 var ErrEmptyByteSlice = errors.New("empty byte array")
 
@@ -958,24 +914,6 @@ const (
 	CovarCols = 16
 )
 
-func (c CovarFlags) String() string {
-	switch c {
-	case CovarScrambled:
-		return "covar-scrambled"
-	case CovarNormal:
-		return "covar-normal"
-	case CovarUseAvg:
-		return "covar-use-avg"
-	case CovarScale:
-		return "covar-scale"
-	case CovarRows:
-		return "covar-rows"
-	case CovarCols:
-		return "covar-cols"
-	}
-	return ""
-}
-
 // CalcCovarMatrix calculates the covariance matrix of a set of vectors.
 //
 // For further details, please see:
@@ -1094,26 +1032,6 @@ const (
 	// DctRows performs a forward or inverse dct transform of every individual row of the input matrix.
 	DctRows = DftRows
 )
-
-func (c DftFlags) String() string {
-	switch c {
-	case DftForward:
-		return "dft-forward"
-	case DftInverse:
-		return "dft-inverse"
-	case DftScale:
-		return "dft-scale"
-	case DftRows:
-		return "dft-rows"
-	case DftComplexOutput:
-		return "dft-complex-output"
-	case DftRealOutput:
-		return "dft-real-output"
-	case DftComplexInput:
-		return "dft-complex-input"
-	}
-	return ""
-}
 
 // DCT performs a forward or inverse discrete Cosine transform of 1D or 2D array.
 //
@@ -1261,18 +1179,6 @@ const (
 	Rotate90CounterClockwise = 2
 )
 
-func (c RotateFlag) String() string {
-	switch c {
-	case Rotate90Clockwise:
-		return "rotate-90-clockwise"
-	case Rotate180Clockwise:
-		return "rotate-180-clockwise"
-	case Rotate90CounterClockwise:
-		return "rotate-90-counter-clockwise"
-	}
-	return ""
-}
-
 // Rotate rotates a 2D array in multiples of 90 degrees
 //
 // For further details, please see:
@@ -1366,18 +1272,6 @@ const (
 	// instead of computing them from the initial centers. For the second and further attempts, use the random or semi-random     // centers. Use one of KMEANS_*_CENTERS flag to specify the exact method.
 	KMeansUseInitialLabels = 2
 )
-
-func (c KMeansFlags) String() string {
-	switch c {
-	case KMeansRandomCenters:
-		return "kmeans-random-centers"
-	case KMeansPPCenters:
-		return "kmeans-pp-centers"
-	case KMeansUseInitialLabels:
-		return "kmeans-use-initial-labels"
-	}
-	return ""
-}
 
 // KMeans finds centers of clusters and groups input samples around the clusters.
 //
@@ -1553,28 +1447,6 @@ const (
 	NormMinMax = 32
 )
 
-func (c NormType) String() string {
-	switch c {
-	case NormInf:
-		return "norm-inf"
-	case NormL1:
-		return "norm-l1"
-	case NormL2:
-		return "norm-l2"
-	case NormL2Sqr:
-		return "norm-l2-sqr"
-	case NormHamming:
-		return "norm-hamming"
-	case NormHamming2:
-		return "norm-hamming2"
-	case NormRelative:
-		return "norm-relative"
-	case NormMinMax:
-		return "norm-minmax"
-	}
-	return ""
-}
-
 // Normalize normalizes the norm or value range of an array.
 //
 // For further details, please see:
@@ -1621,16 +1493,6 @@ const (
 	EPS = 2
 )
 
-func (c TermCriteriaType) String() string {
-	switch c {
-	case Count:
-		return "count"
-	case EPS:
-		return "eps"
-	}
-	return ""
-}
-
 type SolveDecompositionFlags int
 
 const (
@@ -1654,24 +1516,6 @@ const (
 	// 𝚜𝚛𝚌𝟷⋅𝚍𝚜𝚝=𝚜𝚛𝚌𝟸.
 	SolveDecompositionNormal = 5
 )
-
-func (c SolveDecompositionFlags) String() string {
-	switch c {
-	case SolveDecompositionLu:
-		return "solve-decomposition-lu"
-	case SolveDecompositionSvd:
-		return "solve-decomposition-svd"
-	case SolveDecompositionEing:
-		return "solve-decomposition-eing"
-	case SolveDecompositionCholesky:
-		return "solve-decomposition-cholesky"
-	case SolveDecompositionQr:
-		return "solve-decomposition-qr"
-	case SolveDecompositionNormal:
-		return "solve-decomposition-normal"
-	}
-	return ""
-}
 
 // Solve solves one or more linear systems or least-squares problems.
 //
@@ -1715,20 +1559,6 @@ const (
 	// The output is the minimum (column/row-wise) of all rows/columns of the matrix.
 	ReduceMin ReduceTypes = 3
 )
-
-func (c ReduceTypes) String() string {
-	switch c {
-	case ReduceSum:
-		return "reduce-sum"
-	case ReduceAvg:
-		return "reduce-avg"
-	case ReduceMax:
-		return "reduce-max"
-	case ReduceMin:
-		return "reduce-min"
-	}
-	return ""
-}
 
 // Reduce reduces a matrix to a vector.
 //
@@ -1780,18 +1610,6 @@ const (
 	// Each matrix row is sorted in the descending order; this flag and the previous one are also mutually exclusive.
 	SortDescending SortFlags = 16
 )
-
-func (c SortFlags) String() string {
-	switch c {
-	case SortEveryRow:
-		return "sort-every-row"
-	case SortEveryColumn:
-		return "sort-every-column"
-	case SortDescending:
-		return "sort-descending"
-	}
-	return ""
-}
 
 // Sort sorts each row or each column of a matrix.
 //
