+++ cuda/cudawarping.go
@@ -38,26 +38,6 @@ const (
 	InterpolationMax = 7
 )
 
-func (c InterpolationFlags) String() string {
-	switch c {
-	case InterpolationNearestNeighbor:
-		return "interpolation-nearest-neighbor"
-	case InterpolationLinear:
-		return "interpolation-linear"
-	case InterpolationCubic:
-		return "interpolation-cubic"
-	case InterpolationArea:
-		return "interpolation-area"
-	case InterpolationLanczos4:
-		return "interpolation-lanczos4"
-	case InterpolationDefault:
-		return "interpolation-default"
-	case InterpolationMax:
-		return "interpolation-max"
-	}
-	return ""
-}
-
 // BorderType type of border.
 type BorderType int
 
@@ -86,28 +66,6 @@ const (
 	// BorderIsolated border type
 	BorderIsolated = 16
 )
-
-func (c BorderType) String() string {
-	switch c {
-	case BorderConstant:
-		return "border-constant"
-	case BorderReplicate:
-		return "border-replicate"
-	case BorderReflect:
-		return "border-reflect"
-	case BorderWrap:
-		return "border-wrap"
-	case BorderReflect101:
-		return "border-reflect101"
-	case BorderTransparent:
-		return "border-transparent"
-	case BorderDefault:
-		return "border-default"
-	case BorderIsolated:
-		return "border-isolated"
-	}
-	return ""
-}
 
 // Resize resizes an image.
 //
