+++ imgproc.go
@@ -196,24 +196,6 @@ const (
 	HistCmpKlDiv = 5
 )
 
-func (c HistCompMethod) String() string {
-	switch c {
-	case HistCmpCorrel:
-		return "hist-cmp-correl"
-	case HistCmpChiSqr:
-		return "hist-cmp-chi-sqr"
-	case HistCmpIntersect:
-		return "hist-cmp-intersect"
-	case HistCmpBhattacharya:
-		return "hist-cmp-bhattacharya"
-	case HistCmpChiSqrAlt:
-		return "hist-cmp-chi-sqr-alt"
-	case HistCmpKlDiv:
-		return "hist-cmp-kl-div"
-	}
-	return ""
-}
-
 // CompareHist Compares two histograms.
 //
 // For further details, please see:
@@ -321,14 +303,6 @@ const (
 	DistanceLabelPixel
 )
 
-func (c DistanceTransformLabelTypes) String() string {
-	switch c {
-	case DistanceLabelCComp:
-		return "distance-label-ccomp"
-	}
-	return ""
-}
-
 // DistanceTransformMasks are the marsk sizes for distance transform
 type DistanceTransformMasks int
 
@@ -342,14 +316,6 @@ const (
 	// DistanceMaskPrecise is not currently supported
 	DistanceMaskPrecise
 )
-
-func (c DistanceTransformMasks) String() string {
-	switch c {
-	case DistanceMask3:
-		return "distance-mask3"
-	}
-	return ""
-}
 
 // DistanceTransform Calculates the distance to the closest zero pixel for each pixel of the source image.
 //
@@ -396,22 +362,6 @@ const (
 	RetrievalFloodfill = 4
 )
 
-func (c RetrievalMode) String() string {
-	switch c {
-	case RetrievalExternal:
-		return "retrieval-external"
-	case RetrievalList:
-		return "retrieval-list"
-	case RetrievalCComp:
-		return "retrieval-ccomp"
-	case RetrievalTree:
-		return "retrieval-tree"
-	case RetrievalFloodfill:
-		return "retrieval-floodfill"
-	}
-	return ""
-}
-
 // ContourApproximationMode is the mode of the contour approximation algorithm.
 type ContourApproximationMode int
 
@@ -435,20 +385,6 @@ const (
 	// approximation algorithms.
 	ChainApproxTC89KCOS = 4
 )
-
-func (c ContourApproximationMode) String() string {
-	switch c {
-	case ChainApproxNone:
-		return "chain-approx-none"
-	case ChainApproxSimple:
-		return "chain-approx-simple"
-	case ChainApproxTC89L1:
-		return "chain-approx-tc89l1"
-	case ChainApproxTC89KCOS:
-		return "chain-approx-tc89kcos"
-	}
-	return ""
-}
 
 // BoundingRect calculates the up-right bounding rectangle of a point set.
 //
@@ -621,18 +557,6 @@ const (
 	CCL_GRANA = 2
 )
 
-func (c ConnectedComponentsAlgorithmType) String() string {
-	switch c {
-	case CCL_WU:
-		return "ccl-wu"
-	case CCL_DEFAULT:
-		return "ccl-default"
-	case CCL_GRANA:
-		return "ccl-grana"
-	}
-	return ""
-}
-
 // ConnectedComponents computes the connected components labeled image of boolean image.
 //
 // For further details, please see:
@@ -674,24 +598,6 @@ const (
 	CC_STAT_MAX = 5
 )
 
-func (c ConnectedComponentsTypes) String() string {
-	switch c {
-	case CC_STAT_LEFT:
-		return "cc-stat-left"
-	case CC_STAT_TOP:
-		return "cc-stat-top"
-	case CC_STAT_WIDTH:
-		return "cc-stat-width"
-	case CC_STAT_AREA:
-		return "cc-stat-area"
-	case CC_STAT_MAX:
-		return "cc-stat-max"
-	case CC_STAT_HEIGHT:
-		return "cc-stat-height"
-	}
-	return ""
-}
-
 // ConnectedComponentsWithStats computes the connected components labeled image of boolean
 // image and also produces a statistics output for each label.
 //
@@ -732,24 +638,6 @@ const (
 	// TmCcoeffNormed maps to TM_CCOEFF_NORMED
 	TmCcoeffNormed = 5
 )
-
-func (c TemplateMatchMode) String() string {
-	switch c {
-	case TmSqdiff:
-		return "tm-sq-diff"
-	case TmSqdiffNormed:
-		return "tm-sq-diff-normed"
-	case TmCcorr:
-		return "tm-ccorr"
-	case TmCcorrNormed:
-		return "tm-ccorr-normed"
-	case TmCcoeff:
-		return "tm-ccoeff"
-	case TmCcoeffNormed:
-		return "tm-ccoeff-normed"
-	}
-	return ""
-}
 
 // MatchTemplate compares a template against overlapped image regions.
 //
@@ -871,18 +759,6 @@ const (
 	MorphEllipse = 2
 )
 
-func (c MorphShape) String() string {
-	switch c {
-	case MorphRect:
-		return "morph-rect"
-	case MorphCross:
-		return "morph-cross"
-	case MorphEllipse:
-		return "morph-ellispe"
-	}
-	return ""
-}
-
 // GetStructuringElement returns a structuring element of the specified size
 // and shape for morphological operations.
 //
@@ -927,28 +803,6 @@ const (
 	MorphHitmiss = 7
 )
 
-func (c MorphType) String() string {
-	switch c {
-	case MorphErode:
-		return "morph-erode"
-	case MorphDilate:
-		return "morph-dilate"
-	case MorphOpen:
-		return "morph-open"
-	case MorphClose:
-		return "morph-close"
-	case MorphGradient:
-		return "morph-gradient"
-	case MorphTophat:
-		return "morph-tophat"
-	case MorphBlackhat:
-		return "morph-blackhat"
-	case MorphHitmiss:
-		return "morph-hitmiss"
-	}
-	return ""
-}
-
 // BorderType type of border.
 type BorderType int
 
@@ -977,26 +831,6 @@ const (
 	// BorderIsolated border type
 	BorderIsolated = 16
 )
-
-func (c BorderType) String() string {
-	switch c {
-	case BorderConstant:
-		return "border-constant"
-	case BorderReplicate:
-		return "border-replicate"
-	case BorderReflect:
-		return "border-reflect"
-	case BorderWrap:
-		return "border-wrap"
-	case BorderTransparent:
-		return "border-transparent"
-	case BorderDefault:
-		return "border-default"
-	case BorderIsolated:
-		return "border-isolated"
-	}
-	return ""
-}
 
 // GaussianBlur blurs an image Mat using a Gaussian filter.
 // The function convolves the src Mat image into the dst Mat using
@@ -1126,20 +960,6 @@ const (
 	GCEvalFreezeModel = 3
 )
 
-func (c GrabCutMode) String() string {
-	switch c {
-	case GCInitWithRect:
-		return "gc-init-with-rect"
-	case GCInitWithMask:
-		return "gc-init-with-mask"
-	case GCEval:
-		return "gc-eval"
-	case GCEvalFreezeModel:
-		return "gc-eval-freeze-model"
-	}
-	return ""
-}
-
 // Grabcut runs the GrabCut algorithm.
 // The function implements the GrabCut image segmentation algorithm.
 // For further details, please see:
@@ -1174,20 +994,6 @@ const (
 	// 8(1):71–77, 1990.
 	HoughGradient = 3
 )
-
-func (c HoughMode) String() string {
-	switch c {
-	case HoughStandard:
-		return "hough-standard"
-	case HoughProbabilistic:
-		return "hough-probabilistic"
-	case HoughMultiScale:
-		return "hough-multi-scale"
-	case HoughGradient:
-		return "hough-gradient"
-	}
-	return ""
-}
 
 // HoughCircles finds circles in a grayscale image using the Hough transform.
 // The only "method" currently supported is HoughGradient. If you want to pass
@@ -1287,28 +1093,6 @@ const (
 	ThresholdTriangle = 16
 )
 
-func (c ThresholdType) String() string {
-	switch c {
-	case ThresholdBinary:
-		return "threshold-binary"
-	case ThresholdBinaryInv:
-		return "threshold-binary-inv"
-	case ThresholdTrunc:
-		return "threshold-trunc"
-	case ThresholdToZero:
-		return "threshold-to-zero"
-	case ThresholdToZeroInv:
-		return "threshold-to-zero-inv"
-	case ThresholdMask:
-		return "threshold-mask"
-	case ThresholdOtsu:
-		return "threshold-otsu"
-	case ThresholdTriangle:
-		return "threshold-triangle"
-	}
-	return ""
-}
-
 // Threshold applies a fixed-level threshold to each array element.
 //
 // For further details, please see:
@@ -1328,16 +1112,6 @@ const (
 	// AdaptiveThresholdGaussian threshold type
 	AdaptiveThresholdGaussian = 1
 )
-
-func (c AdaptiveThresholdType) String() string {
-	switch c {
-	case AdaptiveThresholdMean:
-		return "adaptative-threshold-mean"
-	case AdaptiveThresholdGaussian:
-		return "adaptative-threshold-gaussian"
-	}
-	return ""
-}
 
 // AdaptiveThreshold applies a fixed-level threshold to each array element.
 //
@@ -1545,30 +1319,6 @@ const (
 	FontItalic = 16
 )
 
-func (c HersheyFont) String() string {
-	switch c {
-	case FontHersheySimplex:
-		return "font-hershey-simplex"
-	case FontHersheyPlain:
-		return "font-hershey-plain"
-	case FontHersheyDuplex:
-		return "font-hershey-duplex"
-	case FontHersheyComplex:
-		return "font-hershey-complex"
-	case FontHersheyTriplex:
-		return "font-hershey-triplex"
-	case FontHersheyComplexSmall:
-		return "font-hershey-complex-small"
-	case FontHersheyScriptSimplex:
-		return "font-hershey-script-simplex"
-	case FontHersheyScriptComplex:
-		return "font-hershey-scipt-complex"
-	case FontItalic:
-		return "font-italic"
-	}
-	return ""
-}
-
 // LineType are the line libraries included in OpenCV.
 //
 // For more information, see:
@@ -1685,24 +1435,6 @@ const (
 	// InterpolationMax indicates use maximum interpolation.
 	InterpolationMax = 7
 )
-
-func (c InterpolationFlags) String() string {
-	switch c {
-	case InterpolationNearestNeighbor:
-		return "interpolation-nearest-neighbor"
-	case InterpolationLinear:
-		return "interpolation-linear"
-	case InterpolationCubic:
-		return "interpolation-cubic"
-	case InterpolationArea:
-		return "interpolation-area"
-	case InterpolationLanczos4:
-		return "interpolation-lanczos4"
-	case InterpolationMax:
-		return "interpolation-max"
-	}
-	return ""
-}
 
 // Resize resizes an image.
 // It resizes the image src down to or up to the specified size, storing the
@@ -1828,36 +1560,6 @@ const (
 	ColormapHot                   = 11
 	ColormapParula                = 12
 )
-
-func (c ColormapTypes) String() string {
-	switch c {
-	case ColormapAutumn:
-		return "colormap-autumn"
-	case ColormapBone:
-		return "colormap-bone"
-	case ColormapJet:
-		return "colormap-jet"
-	case ColormapWinter:
-		return "colormap-winter"
-	case ColormapRainbow:
-		return "colormap-rainbow"
-	case ColormapOcean:
-		return "colormap-ocean"
-	case ColormapSummer:
-		return "colormap-summer"
-	case ColormapSpring:
-		return "colormap-spring"
-	case ColormapCool:
-		return "colormap-cool"
-	case ColormapHsv:
-		return "colormap-hsv"
-	case ColormapPink:
-		return "colormap-pink"
-	case ColormapParula:
-		return "colormap-parula"
-	}
-	return ""
-}
 
 // ApplyColorMap applies a GNU Octave/MATLAB equivalent colormap on a given image.
 //
@@ -2007,26 +1709,6 @@ const (
 	DistWelsch               = 6
 	DistHuber                = 7
 )
-
-func (c DistanceTypes) String() string {
-	switch c {
-	case DistUser:
-		return "dist-user"
-	case DistL1:
-		return "dist-l1"
-	case DistL2:
-		return "dist-l2"
-	case DistL12:
-		return "dist-l12"
-	case DistFair:
-		return "dist-fair"
-	case DistWelsch:
-		return "dist-welsch"
-	case DistHuber:
-		return "dist-huber"
-	}
-	return ""
-}
 
 // FitLine fits a line to a 2D or 3D point set.
 //
