+++ videoio.go
@@ -152,90 +152,6 @@ const (
 	VideoCaptureAutoFocus = 39
 )
 
-func (c VideoCaptureProperties) String() string {
-	switch c {
-	case VideoCapturePosMsec:
-		return "video-capture-pos-msec"
-	case VideoCapturePosFrames:
-		return "video-capture-pos-frames"
-	case VideoCapturePosAVIRatio:
-		return "video-capture-pos-avi-ratio"
-	case VideoCaptureFrameWidth:
-		return "video-capture-frame-width"
-	case VideoCaptureFrameHeight:
-		return "video-capture-frame-height"
-	case VideoCaptureFPS:
-		return "video-capture-fps"
-	case VideoCaptureFOURCC:
-		return "video-capture-fourcc"
-	case VideoCaptureFrameCount:
-		return "video-capture-frame-count"
-	case VideoCaptureFormat:
-		return "video-capture-format"
-	case VideoCaptureMode:
-		return "video-capture-mode"
-	case VideoCaptureBrightness:
-		return "video-capture-brightness"
-	case VideoCaptureContrast:
-		return "video-capture-contrast"
-	case VideoCaptureSaturation:
-		return "video-capture-saturation"
-	case VideoCaptureHue:
-		return "video-capture-hue"
-	case VideoCaptureGain:
-		return "video-capture-gain"
-	case VideoCaptureExposure:
-		return "video-capture-exposure"
-	case VideoCaptureConvertRGB:
-		return "video-capture-convert-rgb"
-	case VideoCaptureWhiteBalanceBlueU:
-		return "video-capture-white-balanced-blue-u"
-	case VideoCaptureWhiteBalanceRedV:
-		return "video-capture-white-balanced-red-v"
-	case VideoCaptureRectification:
-		return "video-capture-rectification"
-	case VideoCaptureMonochrome:
-		return "video-capture-monochrome"
-	case VideoCaptureSharpness:
-		return "video-capture-sharpness"
-	case VideoCaptureAutoExposure:
-		return "video-capture-auto-exposure"
-	case VideoCaptureGamma:
-		return "video-capture-gamma"
-	case VideoCaptureTemperature:
-		return "video-capture-temperature"
-	case VideoCaptureTrigger:
-		return "video-capture-trigger"
-	case VideoCaptureTriggerDelay:
-		return "video-capture-trigger-delay"
-	case VideoCaptureZoom:
-		return "video-capture-zoom"
-	case VideoCaptureFocus:
-		return "video-capture-focus"
-	case VideoCaptureGUID:
-		return "video-capture-guid"
-	case VideoCaptureISOSpeed:
-		return "video-capture-iso-speed"
-	case VideoCaptureBacklight:
-		return "video-capture-backlight"
-	case VideoCapturePan:
-		return "video-capture-pan"
-	case VideoCaptureTilt:
-		return "video-capture-tilt"
-	case VideoCaptureRoll:
-		return "video-capture-roll"
-	case VideoCaptureIris:
-		return "video-capture-iris"
-	case VideoCaptureSettings:
-		return "video-capture-settings"
-	case VideoCaptureBufferSize:
-		return "video-capture-buffer-size"
-	case VideoCaptureAutoFocus:
-		return "video-capture-auto-focus"
-	}
-	return ""
-}
-
 // VideoCapture is a wrapper around the OpenCV VideoCapture class.
 //
 // For further details, please see:
